+++ libs/membership/membership-common/src/main/kotlin/net/corda/membership/lib/approval/RegistrationRulesEngine.kt
@@ -1,5 +1,7 @@
 package net.corda.membership.lib.approval
 
+import net.corda.membership.lib.toMap
+import net.corda.membership.lib.toWire
 import net.corda.v5.membership.MemberInfo
 
 /**
@@ -10,4 +12,18 @@ interface RegistrationRulesEngine {
     val rules: Collection<RegistrationRule>
 
     fun requiresManualApproval(proposedMemberInfo: MemberInfo, activeMemberInfo: MemberInfo?): Boolean
+
+    class Impl(override val rules: Collection<RegistrationRule>) : RegistrationRulesEngine {
+
+        override fun requiresManualApproval(proposedMemberInfo: MemberInfo, activeMemberInfo: MemberInfo?): Boolean {
+            val proposedMemberInfoMap = proposedMemberInfo.memberProvidedContext.toWire().toMap()
+            val memberInfoDiff = activeMemberInfo?.let {
+                proposedMemberInfoMap.keys - it.memberProvidedContext.toWire().toMap().keys
+            } ?: proposedMemberInfoMap.keys
+
+            return rules.any {
+                it.evaluate(memberInfoDiff)
+            }
+        }
+    }
 }
